move Sink and Stream impls to mod channel
